#263 cleaned up
